update: defined product interface and moved product id generator function to common/utils/generateProductId
