+++ backend/src/interfaces/Product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    product_id: string;
+    name: string;
+    image: string;
+    manufacturer: string;
+    manufactureDate: string;
+    expiryDate: string;
+}
